aa2

+++ src/app/views/admin/tables/tables.component.ts
@@ -4,6 +4,7 @@ import { Subscription } from 'rxjs';
 
 import { AdminService } from 'src/app/admin.service';
 import { CreateSubscription } from 'src/app/create-subscription.model';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -40,12 +41,28 @@ ids!: Subscription;
     this.adminService.createSubscription(this.createSubscriptionDto, this.authToken).subscribe(
       response => {
         console.log(response); // Gérez la réponse de manière appropriée
+        Swal.fire({
+          title: 'Succès',
+          text: 'L\'abonnement de sponsoring  a été créée avec succès.',
+          icon: 'success',
+          timer: 1500,
+          showConfirmButton: false
+        });
+        this.getallabonnement();
       },
       error => {
         console.error(error); // Gérez les erreurs de manière appropriée
+        Swal.fire({
+          title: 'Erreur',
+          text: 'Une erreur est survenue lors de la création de la souscription.',
+          icon: 'error',
+          timer: 1500,
+          showConfirmButton: false
+        });
       }
     );
   }
+  
 
   subscriptions: Partial<CreateSubscription>[] = [];
 getallabonnement(){ 
@@ -125,29 +142,67 @@ updateSubscription() {
       (response) => {
         console.log('Subscription updated:', response);
         this.closePopup();
+        // Afficher une alerte de succès
+        Swal.fire({
+          icon: 'success',
+          title: 'Abonnement mis à jour!',
+          showConfirmButton: false,
+          timer: 1500 // Fermer l'alerte après 1.5 secondes
+        });
         // Mettez à jour la liste des abonnements ou effectuez d'autres actions nécessaires
         this.getallabonnement(); // Mettez à jour la liste après la modification
         this.selectedSubscription = undefined; // Effacez la souscription sélectionnée
       },
       (error) => {
         console.error('Error updating subscription:', error);
+        // Afficher une alerte d'erreur
+        Swal.fire({
+          icon: 'error',
+          title: 'Erreur lors de la mise à jour de l\'abonnement',
+          text: 'Veuillez réessayer.',
+        });
       }
     );
 }
 
 deleteSubscription(ids: number): void {
-  this.adminService.deleteSubscription(ids).subscribe(
-    () => {
-      console.log('Abonnement supprimé avec succès');
-      this.getallabonnement();
-      // Mettre à jour la liste des abonnements ou effectuer d'autres actions après la suppression réussie
-    },
-    (error) => {
-      console.error('Erreur lors de la suppression de l\'abonnement : ', error);
-      // Gérer l'erreur comme vous le souhaitez
+  Swal.fire({
+    title: 'Êtes-vous sûr de supprimer cette abonnement ?',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Oui, supprimer!',
+    cancelButtonText: 'Annuler'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.adminService.deleteSubscription(ids).subscribe(
+        () => {
+          console.log('Abonnement supprimé avec succès');
+          this.getallabonnement(); // Mettre à jour la liste des abonnements
+          Swal.fire({
+            title: 'Supprimé!',
+            text: 'L\'abonnement a été supprimé.',
+            icon: 'success',
+            timer: 1500,
+            showConfirmButton: false
+          });
+        },
+        (error) => {
+          console.error('Erreur lors de la suppression de l\'abonnement : ', error);
+          Swal.fire({
+            title: 'Erreur!',
+            text: 'Une erreur est survenue lors de la suppression de l\'abonnement.',
+            icon: 'error',
+            timer: 1500,
+            showConfirmButton: false
+          });
+        }
+      );
     }
-  );
+  });
 }
+
 }
